+++ macgyver-plugin-ci-jenkins/src/main/java/io/macgyver/plugin/ci/jenkins/JenkinsClient.java
@@ -18,6 +18,8 @@ import java.util.Map;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
+import io.macgyver.okrest3.OkRestTarget;
+
 public interface JenkinsClient {
 
 
@@ -58,6 +60,6 @@ public interface JenkinsClient {
 	void quietDown();
 	void cancelQuietDown();
 	
-	
+	OkRestTarget getOkRestTarget();
 	
 }